refactor: 투두 폴더 구조 개선 및 토스트 컴포넌트화
